NoMoreElement; HyperLogLog (not tested); Count (not tested) Signed-off-by: Michel Blancard <michel.blancard@polytechnique.org>

+++ src/Element.java
@@ -16,7 +16,7 @@ public class Element {
 		return Content;
 	}
 	
-	public long GetHash () {
+	public int GetHash () {
 		return ContentHash; 
 	}
